Renamed "interfaces" folder and .rs file to "channels"
