+++ src/channels/serial/board_interface.rs
@@ -1,0 +1,9 @@
+use crate::channels::traits::BoardChannel;
+use crate::channels::SerialChannel;
+use crate::errors::TockloaderError;
+
+impl BoardChannel for SerialChannel {
+    fn open(&mut self) -> Result<(), TockloaderError> {
+        todo!()
+    }
+}
