+++ src/main.rs
@@ -1,6 +1,6 @@
+mod channels;
 mod cli;
 mod errors;
-mod interfaces;
 use cli::make_cli;
 
 fn main() {
